+++ qiniu/io.c
@@ -16,57 +16,32 @@
 typedef struct _Qiniu_Io_form {
 	struct curl_httppost* formpost;
 	struct curl_httppost* lastptr;
-	char* action;
 } Qiniu_Io_form;
+
+static Qiniu_Io_PutExtra qiniu_defaultExtra = { NULL, NULL, 0, 0 };
 
 static void Qiniu_Io_form_init(
 	Qiniu_Io_form* self, const char* uptoken, const char* key, Qiniu_Io_PutExtra* extra)
 {
-	const char* mimeType = extra->mimeType;
-	const char* customMeta = extra->customMeta;
-	const char* callbackParams = extra->callbackParams;
-
-	char* mimeTypeEncoded;
-	char* entryURI;
-	char* entryURIEncoded;
-	char* customMetaEncoded;
-	char* action;
-	char* action2;
-
+	Qiniu_Io_PutExtraParam* param;
 	struct curl_httppost* formpost = NULL;
 	struct curl_httppost* lastptr = NULL;
 
-	if (mimeType == NULL) {
-		mimeType = "application/octet-stream";
+	curl_formadd(&formpost, &lastptr, CURLFORM_COPYNAME, "token", CURLFORM_COPYCONTENTS, uptoken, CURLFORM_END);
+
+	if (extra == NULL) {
+		extra = &qiniu_defaultExtra;
 	}
-	mimeTypeEncoded = Qiniu_String_Encode(mimeType);
-
-	entryURI = Qiniu_String_Concat3(extra->bucket, ":", key);
-	entryURIEncoded = Qiniu_String_Encode(entryURI);
-	free(entryURI);
-
-	action = Qiniu_String_Concat("/rs-put/", entryURIEncoded, "/mimeType/", mimeTypeEncoded, NULL);
-	free(entryURIEncoded);
-	free(mimeTypeEncoded);
-
-	if (customMeta != NULL && *customMeta != '\0') {
-		customMetaEncoded = Qiniu_String_Encode(customMeta);
-		action2 = Qiniu_String_Concat3(action, "/meta/", customMetaEncoded);
-		free(action);
-		free(customMetaEncoded);
-		action = action2;
+	if (key != NULL) {
+		curl_formadd(&formpost, &lastptr, CURLFORM_COPYNAME, "key", CURLFORM_COPYCONTENTS, key, CURLFORM_END);
 	}
-
-	curl_formadd(&formpost, &lastptr, CURLFORM_COPYNAME, "action", CURLFORM_COPYCONTENTS, action, CURLFORM_END);
-	curl_formadd(&formpost, &lastptr, CURLFORM_COPYNAME, "auth", CURLFORM_COPYCONTENTS, uptoken, CURLFORM_END);
-
-	if (callbackParams != NULL && *callbackParams != '\0') {
-		curl_formadd(&formpost, &lastptr, CURLFORM_COPYNAME, "params", CURLFORM_COPYCONTENTS, callbackParams, CURLFORM_END);
+	for (param = extra->params; param != NULL; param = param->next) {
+		curl_formadd(
+			&formpost, &lastptr, CURLFORM_COPYNAME, param->key, CURLFORM_COPYCONTENTS, param->value, CURLFORM_END);
 	}
 
 	self->formpost = formpost;
 	self->lastptr = lastptr;
-	self->action = action;
 }
 
 /*============================================================================*/
@@ -76,27 +51,24 @@ CURL* Qiniu_Client_reset(Qiniu_Client* self);
 Qiniu_Error Qiniu_callex(CURL* curl, Qiniu_Buffer *resp, Qiniu_Json** ret, Qiniu_Bool simpleError, Qiniu_Buffer *resph);
 
 static Qiniu_Error Qiniu_Io_call(
-	Qiniu_Client* self, Qiniu_Io_PutRet* ret, struct curl_httppost* formpost, char* action)
+	Qiniu_Client* self, Qiniu_Io_PutRet* ret, struct curl_httppost* formpost)
 {
 	Qiniu_Error err;
 
 	CURL* curl = Qiniu_Client_reset(self);
-	char* url = Qiniu_String_Concat2(QINIU_UP_HOST, "/upload");
 	struct curl_slist* headers = curl_slist_append(NULL, "Expect:");
 
-	curl_easy_setopt(curl, CURLOPT_URL, url);
+	curl_easy_setopt(curl, CURLOPT_URL, QINIU_UP_HOST);
 	curl_easy_setopt(curl, CURLOPT_HTTPPOST, formpost);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
 
 	err = Qiniu_callex(curl, &self->b, &self->root, Qiniu_False, &self->respHeader);
 	if (err.code == 200 && ret != NULL) {
 		ret->hash = Qiniu_Json_GetString(self->root, "hash", NULL);
+		ret->key = Qiniu_Json_GetString(self->root, "key", NULL);
 	}
 
 	curl_formfree(formpost);
-	free(action);
-	free(url);
-
 	return err;
 }
 
@@ -110,7 +82,7 @@ Qiniu_Error Qiniu_Io_PutFile(
 	curl_formadd(
 		&form.formpost, &form.lastptr, CURLFORM_COPYNAME, "file", CURLFORM_FILE, localFile, CURLFORM_END);
 
-	return Qiniu_Io_call(self, ret, form.formpost, form.action);
+	return Qiniu_Io_call(self, ret, form.formpost);
 }
 
 Qiniu_Error Qiniu_Io_PutBuffer(
@@ -124,6 +96,6 @@ Qiniu_Error Qiniu_Io_PutBuffer(
 		&form.formpost, &form.lastptr, CURLFORM_COPYNAME, "file",
 		CURLFORM_BUFFER, key, CURLFORM_BUFFERPTR, buf, CURLFORM_BUFFERLENGTH, fsize, CURLFORM_END);
 
-	return Qiniu_Io_call(self, ret, form.formpost, form.action);
+	return Qiniu_Io_call(self, ret, form.formpost);
 }